+++ packages/auth/android/src/main/java/io/invertase/firebase/auth/ReactNativeFirebaseAuthModule.java
@@ -2576,6 +2576,11 @@ class ReactNativeFirebaseAuthModule extends ReactNativeFirebaseModule {
       builder = builder.setDynamicLinkDomain(actionCodeSettings.getString("dynamicLinkDomain"));
     }
 
+    if (actionCodeSettings.hasKey("linkDomain")) {
+      builder =
+          builder.setLinkDomain(Objects.requireNonNull(actionCodeSettings.getString("linkDomain")));
+    }
+
     if (actionCodeSettings.hasKey("android")) {
       ReadableMap android = actionCodeSettings.getMap("android");
       boolean installApp =