+++ packages/firestore/lib/modular/Bytes.d.ts
@@ -1,4 +1,4 @@
-export declare class Bytes {
+export declare class Bytes extends FirestoreBlob {
   static fromBase64String(base64: string): Bytes;
 
   static fromUint8Array(array: Uint8Array): Bytes;